rename `entrypoints.ts` to `execute.ts`
