+++ src/execution/__tests__/errorPropagation-test.ts
@@ -8,7 +8,7 @@ import { parse } from '../../language/parser.js';
 
 import { buildSchema } from '../../utilities/buildASTSchema.js';
 
-import { execute } from '../entrypoints.js';
+import { execute } from '../execute.js';
 import type { ExecutionResult } from '../Executor.js';
 
 const syncError = new Error('bar');
