+++ src/execution/__tests__/executor-test.ts
@@ -30,7 +30,7 @@ import {
 } from '../../type/scalars.js';
 import { GraphQLSchema } from '../../type/schema.js';
 
-import { execute, executeSync } from '../entrypoints.js';
+import { execute, executeSync } from '../execute.js';
 
 describe('Execute: Handles basic execution tasks', () => {
   it('executes arbitrary code', async () => {
