+++ src/execution/__tests__/lists-test.ts
@@ -18,7 +18,7 @@ import { GraphQLSchema } from '../../type/schema.js';
 
 import { buildSchema } from '../../utilities/buildASTSchema.js';
 
-import { execute, executeSync } from '../entrypoints.js';
+import { execute, executeSync } from '../execute.js';
 import type { ExecutionResult } from '../Executor.js';
 
 describe('Execute: Accepts any iterable as list value', () => {
